Extract loudness analyzer from ImportService
Move the loudness measurement unit — the meter sink, its DecodedSink impl, and
measure_loudness — out of ImportService into a self-contained import::loudness
module. The analyzer takes the broadcast sender it emits progress on as an
argument instead of reaching back through ImportService, so it no longer
depends on the service type; run_import drives it with &self.event_tx. The
loudness-progress emitter moves with it. No public API change.
