+++ bae-core/src/import/mod.rs
@@ -15,6 +15,8 @@ mod image_response;
 // below (re-exported from `types`) compile on mobile.
 #[cfg(not(any(target_os = "ios", target_os = "android")))]
 mod handle;
+#[cfg(any(target_os = "macos", target_os = "linux", target_os = "windows"))]
+mod loudness;
 pub mod musicbrainz_mapper;
 #[cfg(not(any(target_os = "ios", target_os = "android")))]
 mod progress;
